refactor(instance): Separate the mod_loader from the mod.
