+++ src-tauri/src/instance/helpers/mod.rs
@@ -4,7 +4,6 @@ pub mod client_json;
 pub mod game_version;
 pub mod hmcl_patch;
 pub mod misc;
-pub mod mod_loader;
 pub mod modpack;
 pub mod mods;
 pub mod resourcepack;
