+++ src-tauri/src/instance/helpers/mods/loader/mod.rs
@@ -1,0 +1,4 @@
+pub mod common;
+pub mod fabric;
+pub mod forge;
+pub mod neoforge;
