+++ src-tauri/src/instance/helpers/mods/mod.rs
@@ -3,4 +3,5 @@ pub mod fabric;
 pub mod forge;
 pub mod legacy_forge;
 pub mod liteloader;
+pub mod loader;
 pub mod quilt;
